refactor: vertical slice
